Extracted movement list into separate component
