--- frontend/erc20sepolia.json
@@ -1,20 +1,0 @@
-{
-  "tokens": [
-    {
-      "name": "DAI",
-      "symbol": "DAI",
-      "address": "0xf5c142292B85253e4D071812c84f05ec42828fdB"
-    },
-    {
-      "name": "Tether USD",
-      "symbol": "USDT",
-      "address": "0xaA8E23Fb1079EA71e0a56F48a2aA51851D8433D0"
-    },
-    {
-      "name": "USD Coin",
-      "symbol": "USDC",
-      "address": "0x73d219B3881E481394DA6B5008A081d623992200"
-    }
-  ]
-}
-
