+++ frontend/netconfig/blockchain-config.d.ts
@@ -8,12 +8,12 @@ interface ERC20Token {
     symbol: string;
     address: string;
   }
-  
+
 interface BlockchainConfig {
     block_explorer: string;
     erc20_tokens: ERC20Token[];
 }
-  
+
 interface Config {
     sepolia: BlockchainConfig;
     ethereum: BlockchainConfig;
