+++ frontend/src/app/(auth)/(tabs)/index.tsx
@@ -1,53 +1,24 @@
 import { AssetTransfersResult } from "alchemy-sdk";
-import { Link, router } from "expo-router";
-import React, { Key } from "react";
-import { ColorValue, ScrollView, View } from "react-native";
+import { Link } from "expo-router";
+import React from "react";
+import { ScrollView, View } from "react-native";
 import {
   ActivityIndicator,
   Avatar,
   Card,
   FAB,
-  List,
   Text,
 } from "react-native-paper";
 import { useBalance } from "../../../hooks/useBalance";
 import { useRecentTransactions } from "../../../hooks/useRecentTransactions";
 import styles from "../../../styles/styles";
 import { useAuthContext } from "../../../providers/AuthProvider";
-
-const formatTxAddress = (address: string) => {
-  return `${address.slice(0, 10)}...${address.slice(30)}`;
-};
-
-const EthereumTransactionItem = (
-  transfer: AssetTransfersResult,
-  index: Key,
-  color: ColorValue
-) => (
-  <List.Item
-    title={`From: ${formatTxAddress(transfer.from)}`}
-    titleStyle={{ fontSize: 12 }}
-    description={
-      `In block: ${transfer.blockNum}` +
-      "\n" +
-      `Tx ID: ${formatTxAddress(transfer.uniqueId)}`
-    }
-    descriptionStyle={{ fontSize: 12 }}
-    key={index}
-    onPress={() =>
-      router.push({
-        pathname: `transaction`,
-        params: { txHash: transfer.hash },
-      })
-    }
-    right={() => <Text style={{ color: color }}>{transfer.value}</Text>}
-  />
-);
+import MovementsList from "../../../components/movementsList";
 
 const HomeScreen: React.FC = () => {
   const balance = useBalance();
   const { fromTransfers, toTransfers } = useRecentTransactions();
-  const { user } = useAuthContext()
+  const { user } = useAuthContext();
 
   return (
     <ScrollView>
@@ -56,6 +27,7 @@ const HomeScreen: React.FC = () => {
           <Avatar.Icon style={styles.userSettings} size={30} icon="account" />
           <Text variant="titleLarge">{`Hola ${user?.email}`}</Text>
         </View>
+
         <Card style={styles.movementsCard} mode="contained">
           <Card.Content>
             <Text variant="titleLarge">Balance</Text>
@@ -77,31 +49,14 @@ const HomeScreen: React.FC = () => {
           <FAB size="small" icon="cash-minus" variant="tertiary" />
           <FAB size="small" icon="account-cash-outline" variant="tertiary" />
         </View>
+
         <Card style={styles.movementsCard} mode="elevated">
           <Card.Content>
             <Card.Title title="Movements" />
-            <List.Section>
-              <List.Subheader key={"Received"}>Received</List.Subheader>
-              {!toTransfers && <ActivityIndicator />}
-              {toTransfers &&
-                toTransfers.map((transfer, index) =>
-                  EthereumTransactionItem(
-                    transfer,
-                    `to_transfer_${index}`,
-                    "green"
-                  )
-                )}
-              <List.Subheader key={"Sent"}>Sent</List.Subheader>
-              {!fromTransfers && <ActivityIndicator />}
-              {fromTransfers &&
-                fromTransfers.map((transfer, index) =>
-                  EthereumTransactionItem(
-                    transfer,
-                    `from_transfer_${index}`,
-                    "red"
-                  )
-                )}
-            </List.Section>
+            <MovementsList
+              toTransfers={toTransfers}
+              fromTransfers={fromTransfers}
+            />
           </Card.Content>
         </Card>
       </View>
